refactor: extract poll dimension sync status into a global helper function
