+++ src/app/core/services/common/helper.service.ts
@@ -13,6 +13,12 @@ import { brandingConfig } from 'src/app/branding/branding-config';
 import { SentenceCasePipe } from 'src/app/shared/pipes/sentence-case.pipe';
 import { TitleCasePipe } from '@angular/common';
 import { DefaultDestinationAttribute, DestinationAttribute } from '../../models/db/destination-attribute.model';
+import { Observable, interval, take } from 'rxjs';
+
+type PollDimensionsSyncStatusParams = {
+  onPollingComplete: () => void
+  getWorkspacesObserver: () => Observable<{destination_synced_at: any}[]>
+}
 
 @Injectable({
   providedIn: 'root'
@@ -286,4 +292,39 @@ export class HelperService {
     options.sort((a, b) => (a.name || '').localeCompare(b.name || ''));
   }
 
+
+  /**
+   * Checks every `DIMENSIONS_POLLING_INTERVAL` ms whether dimensions have been refreshed
+   * for a maximum of `DIMENSIONS_POLLING_TIMEOUT` ms.
+   *
+   * Terminates either when dimensions have been refreshed, when refresh dimensions fails,
+   * or when polling times out - whichever occurs first.
+   */
+  pollDimensionsSyncStatus({onPollingComplete, getWorkspacesObserver}: PollDimensionsSyncStatusParams) {
+
+    /** Time (in ms) to wait before each request while polling for refresh dimensions status */
+    const DIMENSIONS_POLLING_INTERVAL = 3000;
+
+    /** Maximum time (in ms) to wait before terminating the poll for refresh dimensions status */
+    const DIMENSIONS_POLLING_TIMEOUT = 60_000;
+
+    const ticks = Math.floor(DIMENSIONS_POLLING_TIMEOUT / DIMENSIONS_POLLING_INTERVAL);
+
+    const pollingSubscription = interval(DIMENSIONS_POLLING_INTERVAL).pipe(take(ticks)).subscribe(
+      {
+        next: (_tick) => {
+          getWorkspacesObserver().subscribe(workspaces => {
+            const {destination_synced_at} = workspaces[0];
+            if (destination_synced_at !== null) {
+              onPollingComplete();
+              pollingSubscription.unsubscribe();
+            }
+          });
+        },
+
+        complete: () => {
+          onPollingComplete();
+        }
+    });
+  }
 }
