+++ src/app/shared/components/si/core/intacct-location-entity/intacct-location-entity.component.ts
@@ -14,12 +14,7 @@ import { SiMappingsService } from 'src/app/core/services/si/si-core/si-mappings.
 import { IntacctDestinationAttribute } from 'src/app/core/models/intacct/db/destination-attribute.model';
 import { brandingConfig, brandingContent, brandingFeatureConfig, brandingKbArticles } from 'src/app/branding/branding-config';
 import { interval, take } from 'rxjs';
-
-/** Time (in ms) to wait before each request while polling for refresh dimensions status */
-const DIMENSIONS_POLLING_INTERVAL = 3000;
-
-/** Maximum time (in ms) to wait before terminating the poll for refresh dimensions status */
-const DIMENSIONS_POLLING_TIMEOUT = 60_000;
+import { HelperService } from 'src/app/core/services/common/helper.service';
 
 @Component({
   selector: 'app-intacct-location-entity',
@@ -67,7 +62,8 @@ export class IntacctLocationEntityComponent implements OnInit {
     private router: Router,
     private workspaceService: SiWorkspaceService,
     private toastService: IntegrationsToastService,
-    private trackingService: TrackingService
+    private trackingService: TrackingService,
+    private helperService: HelperService
   ) { }
 
   patchFormValue(event: any): void {
@@ -133,41 +129,13 @@ export class IntacctLocationEntityComponent implements OnInit {
     this.isRefreshDimensionInProgress = true;
 
     this.mappingsService.refreshSageIntacctDimensions().subscribe();
-    this.pollDimensionsSyncStatus({
+
+    const fyleOrgId = this.storageService.get('org').fyle_org_id;
+    this.helperService.pollDimensionsSyncStatus({
       onPollingComplete: () => {
         this.setOnboardingStateAndRedirect(locationEntityMappingPayload);
-      }
-    });
-
-  }
-
-  /**
-   * Checks every `DIMENSIONS_POLLING_INTERVAL` ms whether dimensions have been refreshed
-   * for a maximum of `DIMENSIONS_POLLING_TIMEOUT` ms.
-   *
-   * Terminates either when dimensions have been refreshed, when refresh dimensions fails,
-   * or when polling times out - whichever occurs first.
-   */
-  private pollDimensionsSyncStatus({onPollingComplete}: {onPollingComplete: () => void}) {
-    const ticks = Math.floor(DIMENSIONS_POLLING_TIMEOUT / DIMENSIONS_POLLING_INTERVAL);
-    const fyleOrgId = this.storageService.get('org').fyle_org_id;
-
-    const pollingSubscription = interval(DIMENSIONS_POLLING_INTERVAL).pipe(take(ticks)).subscribe(
-      {
-        next: (_tick) => {
-          this.workspaceService.getWorkspaceWithoutCache(fyleOrgId, true).subscribe(workspaces => {
-            const {destination_synced_at} = workspaces[0];
-
-            if (destination_synced_at !== null) {
-              onPollingComplete();
-              pollingSubscription.unsubscribe();
-            }
-          });
-        },
-
-        complete: () => {
-          onPollingComplete();
-        }
+      },
+      getWorkspacesObserver: () => this.workspaceService.getWorkspaceWithoutCache(fyleOrgId, true)
     });
   }
